coalib: Rename coala_bears_test_module
Rename coala_bears_test_module to test_bears.

Fixes: https://github.com/coala-analyzer/coala/issues/1616
