+++ coalib/tests/test_bears/LineCountTestBear.py
