+++ coalib/tests/test_bears/internal_folder/SpaceConsistencyTestBear.py
